[libtooling] enforce that variant scopes have one argument
Summary:
We implicitely use the convention that `emitSimpleVariant` is used if and only
the variant has no argument. This enforces this discipline, as it will be
useful for biniou later.

This superseeds and removes the previous mechanism that was in place to do
sanity checks on variants.

Test Plan:
  DEBUG=1 CC=clang CXX=clang++ CFLAGS='-std=c++11 -fPIC -fblocks' LDFLAGS="-shared" make -C libtooling test
  DEBUG=1 CC=clang CXX=clang++ CFLAGS='-std=c++11 -fPIC -fblocks' LDFLAGS="-shared" make test

Also, play with jsontest.cpp to add/remove arguments to `VariantScope` and see
that assert fails using the commands above.

Reviewers: akotulski, mathieubaudet

Reviewed By: mathieubaudet

Differential Revision: https://phabricator.fb.com/D2287579

Tasks: 7840026

Signature: t1:2287579:1438268624:a31cc345f64b6c5ea64118e268d74a1b6fb2d975