Add a map from ivars to their corresponding properties as this link is missing in clang
Summary:
When ivars are created automatically as part of a property, there is a link from the property to the ivar,
but not from the ivar to the property, and this is useful to have sometimes, for instance, to get the properties of
the ivar.

Test Plan: Compilation and tests pass.

Reviewers: mathieubaudet, akotulski

Reviewed By: akotulski

Subscribers: martinoluca, shb, mathieubaudet, jjb

Differential Revision: https://phabricator.intern.facebook.com/D3542650

Tasks: 12113010

Signature: t1:3542650:1468316671:a57cff3763e0314c8a82e519def267c4ee742b16