Dump GenericSelectionExpr
Summary:
From the clang docs:

"A generic selection contains an unevaluated controlling expression, followed by one or more generic associations. Each generic association specifies a type name and an expression, or "default" and an expression (in which case it is known as a default generic association). The type and value of the generic selection are identical to those of its result expression, which is defined as the expression in the generic association with a type name that is compatible with the type of the controlling expression, or the expression in the default generic association if no types are compatible. For example:

`_Generic(X, double: 1, float: 2, default: 3)`"

Now we dump the result expression in case it exists.

Test Plan:
Added a test, also checked that it works with Infer, in particular it helps fix a big source of dead store fps.

`make -C libtooling test` passes.

Reviewers: shb, mbouaziz, daivan

Reviewed By: shb

Subscribers: varunarora

Differential Revision: https://phabricator.intern.facebook.com/D7169066

Tasks: T26323788

Signature: 7169066:1520386901:3f870def584abb08cc19b1d65ea10a5f50b789a1