[clang] upgrade to 8.0.0
Summary:
Upgrading the version of clang.

- Some new cases in enums showed up when upgrading to clang 8.
- biggest change is a new way to dump attributes

Previously we dumped attributes by hijacking the printing of attributes. The
trick was to pass it a custom output stream that records each individual output
step as elemenst of a vector. The hope is that each output step corresponds to
roughly one parameter of the attribute.

Of course this didn't really work but since we only look for a handful of
attributes this was good enough.

The printing function we used doesn't seem to be exported anymore in clang 8,
which forces us to implement a less clowny solution based on the attribute
visitor provided by libtooling.

The structure of the visitor is the same as the visitor of the AST itself,
including the same preprocessor tricks for generating the ATD OCaml data
structure type as described in libtooling/ATD_GUIDELINES.md.

Test Plan: tested end to end on linux and osx at the top of the stack

Reviewers: ezgi, mbouaziz, mityal

Reviewed By: mbouaziz

Differential Revision: https://phabricator.intern.facebook.com/D15450899

Tasks: T44636077

Signature: 15450899:1558557588:3d364db7df5380cec81f6b5eb45fe5fc7b3ea19e