[libtooling] export structure of IfStmt
Summary:
The structure of `if` statements has changed (https://reviews.llvm.org/D53607)
and now we cannot just look at the list of sub-statements to figure out how to
translate the `if` because some substatements are now optional so we cannot
know which substatement correspond to which part when just looking at them in
order. So, we need to also output which child corresponds to which part of the
if: the init part (the `int x = foo()` bit of `if (int x = foo(); x == 4) { ...
}`), the conditional variable (`x`) in the previous example, the condition
itself (`x == 4`), and the body of `then` and (optionally now!) `else`.

Test Plan: tested end to end in infer

Reviewers: ezgi, mbouaziz, mityal

Reviewed By: mbouaziz

Differential Revision: https://phabricator.intern.facebook.com/D15450897

Tasks: T44636077

Signature: 15450897:1558558181:e18d9237970be0d9a0e3f7c4b3ab376ce7ab0566