Dump ObjCBridgedCastExpr
Summary: This was not exported and is needed to translate correctly the construct `__bridge_transfer`.

Test Plan:
Updated the tests in `make -C libtooling test` and `make -C clang-ocaml test`.

Many results changed because I had to add the flag `fbobjc-arc` to the tests since this construct is only available in arc. It also makes sense to use it since all modern ObjC uses it.

I checked in the test that we get the expected tags in the json.

Reviewers: daivan, mtrojer, arr, jul

Reviewed By: jul

Differential Revision: https://phabricator.intern.facebook.com/D21201914

Tasks: T64840540

Signature: 21201914:1587646813:cef65ce7934191ae58eb67073fe8a7ecc6e91d82